% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiationFunctions.R
\name{BeamOnLeafAzimuth}
\alias{BeamOnLeafAzimuth}
\title{calculation of radiation intensity on a sloped plane from sunAngle and Radiation intensity of
a flat surface}
\usage{
BeamOnLeafAzimuth(
  BeamFlat,
  SunAngle,
  LeafAngle,
  SunAzimuthAngle,
  LeafAzimuthAngle
)
}
\arguments{
\item{BeamFlat}{radiation intensity on a flat surface \link{W/m²}}

\item{SunAngle}{angle of the sun \link{degrees}}

\item{LeafAngle}{angle of the leaf \link{degrees}}

\item{SunAzimuthAngle}{Azimuth angle of the sun \link{degrees}}

\item{LeafAzimuthAngle}{Azimuth angle of the leaf \link{degrees}}
}
\value{
radiation intensity on a sloped leaf plane \link{W/m²}
}
\description{
calculation of radiation intensity on a sloped plane from sunAngle and Radiation intensity of
a flat surface
}
