% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiationFunctions.R
\name{chiFromMeanAngle}
\alias{chiFromMeanAngle}
\title{calculation of chi ratio from mean leaf angle}
\usage{
chiFromMeanAngle(leaf_angle)
}
\arguments{
\item{leaf_angle}{mean leaf angle \link{degrees}}
}
\value{
chi ratio
}
\description{
calculation of chi ratio from mean leaf angle
}
