% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiationFunctions.R
\name{Elevation}
\alias{Elevation}
\title{Title sun elevation angle}
\usage{
Elevation(solar.hour, latitude, day.of.year)
}
\arguments{
\item{solar.hour}{Solar hour \link{0-24}}

\item{latitude}{Latitude \link{degrees}}

\item{day.of.year}{Day of the year \link{1-365}}
}
\value{
sun elevation angle \link{degrees}
}
\description{
Title sun elevation angle
}
