% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiationFunctions.R
\name{extinctionCoefficientEllipsoidal}
\alias{extinctionCoefficientEllipsoidal}
\title{Calculation of extinction coefficient for ellipsoidal leaf angle distribution}
\usage{
extinctionCoefficientEllipsoidal(chi_ratio, sun_angle)
}
\arguments{
\item{chi_ratio}{ratio of length to width of the ellipsoid (chi)}

\item{sun_angle}{angle of the sun \link{degrees}}
}
\value{
extinction coefficient for ellipsoidal leaf angle distribution
}
\description{
Calculation of extinction coefficient for ellipsoidal leaf angle distribution
}
