% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiationFunctions.R
\name{f_daylength}
\alias{f_daylength}
\title{Title}
\usage{
f_daylength(DAY, LAT)
}
\arguments{
\item{DAY}{Day of the year \link{1-365}}

\item{LAT}{Latitude \link{degrees}}
}
\value{
daylength \link{h}
}
\description{
Title
}
