% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiationFunctions.R
\name{f_daylengthp}
\alias{f_daylengthp}
\title{Title}
\usage{
f_daylengthp(DAY, LAT)
}
\arguments{
\item{DAY}{day of the year \link{1-365}}

\item{LAT}{latitude \link{degrees}}
}
\value{
photoperiodic daylength \link{h}
}
\description{
Title
}
