% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiationFunctions.R
\name{f_dsinbeta}
\alias{f_dsinbeta}
\title{dsinbeta function}
\usage{
f_dsinbeta(LAT, DAY, DAYL)
}
\arguments{
\item{LAT}{latitude \link{degrees}}

\item{DAY}{day of the year \link{1-365}}

\item{DAYL}{daylength \link{h}}
}
\value{
dsinbeta
}
\description{
dsinbeta function
}
