% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiationFunctions.R
\name{f_frdif_day}
\alias{f_frdif_day}
\title{fraction of diffuse radiation based on daily observations}
\usage{
f_frdif_day(ATMTR)
}
\arguments{
\item{ATMTR}{atmospheric transmission}
}
\value{
fraction of diffuse radiation \link{-}
}
\description{
fraction of diffuse radiation based on daily observations
}
