% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiationFunctions.R, R/SucrosFunctions.R
\name{f_frdif_hour}
\alias{f_frdif_hour}
\title{Estimation of fraction of diffuse radiation on hourly basis}
\usage{
f_frdif_hour(Sg_S0, lat, day, hour)

f_frdif_hour(Sg_S0, lat, day, hour)
}
\arguments{
\item{Sg_S0}{relation of measured to maximum solar radiation}

\item{lat}{latitude \link{°}}

\item{day}{day of year}

\item{hour}{hour of day}
}
\value{
fraction of diffuse radiation \link{-}

frdif fraction of diffuse radiation
}
\description{
Estimation of fraction of diffuse radiation on hourly basis

Title
}
