% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiationFunctions.R
\name{f_sinbeta}
\alias{f_sinbeta}
\title{Title}
\usage{
f_sinbeta(lat, day, hour)
}
\arguments{
\item{lat}{Latitude \link{degrees}}

\item{day}{day of the year \link{1-365}}

\item{hour}{hour of the day \link{0-24}}
}
\value{
sin beta (sun elevation)
}
\description{
Title
}
