% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SucrosFunctions.R
\name{f_SucrosTransmission}
\alias{f_SucrosTransmission}
\title{Title f_SucrosTransmission
functions for calculation of radiation transmission
according to the sucros model}
\usage{
f_SucrosTransmission(lai, sg, lat, day, hour)
}
\arguments{
\item{lai}{leaf area index}

\item{sg}{global radiation \link{J/m2/s}}

\item{lat}{latitude \link{°}}

\item{day}{day of year}

\item{hour}{hour of day}
}
\value{
transmission fraction of transmitted radiation
}
\description{
Title f_SucrosTransmission
functions for calculation of radiation transmission
according to the sucros model
}
