% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiationFunctions.R
\name{k_cylindrical}
\alias{k_cylindrical}
\title{Extinction coefficient for cylindrical leaf angle distribution}
\usage{
k_cylindrical(sun_angle)
}
\arguments{
\item{sun_angle}{angle of the sun \link{degrees}}
}
\value{
extinction coefficient for cylindrical leaf angle distribution
}
\description{
Extinction coefficient for cylindrical leaf angle distribution
}
