% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiationFunctions.R
\name{k_spherical}
\alias{k_spherical}
\title{Extinction coefficient for spherical leaf angle distribution}
\usage{
k_spherical(sun_angle)
}
\arguments{
\item{sun_angle}{angle of the sun \link{degrees}}
}
\value{
extinction coefficient for spherical leaf angle distribution
}
\description{
Extinction coefficient for spherical leaf angle distribution
}
