% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiationFunctions.R
\name{KdifEllipsoidal}
\alias{KdifEllipsoidal}
\title{Calculation of extinction coefficient for diffuse radiation
with ellipsoidal leaf angle distribution}
\usage{
KdifEllipsoidal(lai, chi_ratio)
}
\arguments{
\item{lai}{leaf area index \link{dimensionless}}

\item{chi_ratio}{ratio of length to width of the ellipsoid (chi)}
}
\value{
extinction coefficient for diffuse radiation
}
\description{
Calculation of extinction coefficient for diffuse radiation
with ellipsoidal leaf angle distribution
}
