% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiationFunctions.R
\name{leafAngleFractionEllipsoidal}
\alias{leafAngleFractionEllipsoidal}
\title{Fraction of leaf angle distribution}
\usage{
leafAngleFractionEllipsoidal(chi_ratio, leaf_angle)
}
\arguments{
\item{chi_ratio}{ratio of length to width of the ellipsoid}

\item{leaf_angle}{angle of the leaf \link{degrees}}
}
\value{
fraction of leaf angle distribution
}
\description{
Fraction of leaf angle distribution
}
