% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiationFunctions.R
\name{MeanAngleFromChi}
\alias{MeanAngleFromChi}
\title{Function to calculate mean leaf angle from chi ratio}
\usage{
MeanAngleFromChi(chi)
}
\arguments{
\item{chi}{ratio ratio of length to width of the ellipsoid (chi)}
}
\value{
mean leaf angle \link{degrees}
}
\description{
Function to calculate mean leaf angle from chi ratio
}
