% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhotoFunctions.R
\name{NegExponential}
\alias{NegExponential}
\title{Rectangular hyperbola function for photosynthesis}
\usage{
NegExponential(Pmax, alpha, I)
}
\arguments{
\item{Pmax}{light saturated photosynthesis rate \link{µmol CO2 m^-2 s^-1}}

\item{alpha}{quantum yield, initial slope \link{µmol CO2 µmol^-1 photons}}

\item{I}{photosynthetic photon flux density, PPFD \link{µmol photons m^-2 s^-1}}
}
\value{
single leaf photosynthesis rate \link{µmol CO2 m^-2 s^-1}
}
\description{
Rectangular hyperbola function for photosynthesis
}
