% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhotoFunctions.R
\name{NonRectHypR}
\alias{NonRectHypR}
\title{Non-rectangular hyperbola function for photosynthesis with respiration}
\usage{
NonRectHypR(Pmax, alpha, r, theta, I)
}
\arguments{
\item{Pmax}{light saturated photosynthesis rate \link{µmol CO2 m^-2 s^-1}}

\item{alpha}{quantum yield, initial slope \link{µmol CO2 µmol^-1 photons}}

\item{r}{dark respiration rate \link{µmol CO2 m^-2 s^-1}}

\item{theta}{curvature parameter \link{dimensionless}}

\item{I}{photosynthetic photon flux density, PPFD \link{µmol photons m^-2 s^-1}}
}
\value{
single leaf net photosynthesis rate \link{µmol CO2 m^-2 s^-1}
}
\description{
Non-rectangular hyperbola function for photosynthesis with respiration
}
