% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiationFunctions.R
\name{r_net}
\alias{r_net}
\title{Estimation of net radiation}
\usage{
r_net(albedo = 0.77, R_s, R_sclear, Airtemp = 20, e_a)
}
\arguments{
\item{albedo}{albedo of the surface \link{-}}

\item{R_s}{global radiation \link{W/m²}}

\item{R_sclear}{clear sky radiation \link{W/m²}}

\item{Airtemp}{air temperature \link{°C}}

\item{e_a}{actual vapor pressure \link{hPa}}
}
\value{
net radiation \link{W/m²}
}
\description{
Estimation of net radiation
}
