% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhotoFunctions.R
\name{RectHypIntPmaxdec}
\alias{RectHypIntPmaxdec}
\title{Canopy integral of rectangular hyperbola function for photosynthesis with decrease Pmax of Pmax
and respiration}
\usage{
RectHypIntPmaxdec(LAI, Pmax, alpha, k, I)
}
\arguments{
\item{LAI}{leaf area index \link{dimensionless}}

\item{Pmax}{light saturated photosynthesis rate \link{µmol CO2 m^-2 s^-1}}

\item{alpha}{quantum yield, initial slope \link{µmol CO2 µmol^-1 photons}}

\item{k}{extinction coefficient \link{dimensionless}}

\item{I}{photosynthetic photon flux density, PPFD \link{µmol photons m^-2 s^-1}}
}
\value{
canopy-integrated photosynthesis rate \link{µmol CO2 m^-2 s^-1}
}
\description{
Canopy integral of rectangular hyperbola function for photosynthesis with decrease Pmax of Pmax
and respiration
}
