% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhotoFunctions.R
\name{RectHypIntR}
\alias{RectHypIntR}
\title{Canopy Integral of rectangular hyperbola function for photosynthesis with constant Pmax and respiration}
\usage{
RectHypIntR(LAI, Pmax, alpha, r, k, I)
}
\arguments{
\item{LAI}{leaf area index \link{dimensionless}}

\item{Pmax}{light saturated photosynthesis rate \link{µmol CO2 m^-2 s^-1}}

\item{alpha}{quantum yield, initial slope \link{µmol CO2 µmol^-1 photons}}

\item{r}{dark respiration rate per unit leaf area \link{µmol CO2 m^-2 s^-1}}

\item{k}{extinction coefficient \link{dimensionless}}

\item{I}{photosynthetic photon flux density, PPFD \link{µmol photons m^-2 s^-1}}
}
\value{
canopy-integrated net photosynthesis rate \link{µmol CO2 m^-2 s^-1}
}
\description{
Canopy Integral of rectangular hyperbola function for photosynthesis with constant Pmax and respiration
}
