% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhotoFunctions.R
\name{sinusf}
\alias{sinusf}
\title{Utility function for transfer of a daily average to a 16 hour sinoidal function}
\usage{
sinusf(hour = 12)
}
\arguments{
\item{hour}{hour of the day \link{0-24}}
}
\value{
the actual value of the sinusoidal function for that hour
}
\description{
Utility function for transfer of a daily average to a 16 hour sinoidal function
}
