% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SucrosFunctions.R
\name{trapez_f}
\alias{trapez_f}
\title{trapezoid function}
\usage{
trapez_f(x, x0 = 0, x1 = 10, x2 = 18, x3 = 35, fmin = 0, fmax = 1)
}
\arguments{
\item{x}{independent variable}

\item{x0}{lower limit where Y = fmin}

\item{x1}{lower limit where Y = fmax}

\item{x2}{upper limit where Y = fmax}

\item{x3}{upper limit where Y = fmin}

\item{fmin}{lowest value of Y}

\item{fmax}{highest value of Y}
}
\value{
interpolated y value of the trapezoid function
}
\description{
trapezoid function
}
